+++ docs/source/swagger-yaml-to-html.py
@@ -16,7 +16,10 @@ Usage:
     python swagger-yaml-to-html.py < /path/to/api.yaml > doc.html
 
 """
-import yaml, json, sys
+import json
+import sys
+
+import yaml
 
 TEMPLATE = """
 <!DOCTYPE html>