+++ globus_action_provider_tools/data_types.py
@@ -1,17 +1,16 @@
 import datetime
 import inspect
+import json
 from enum import Enum
-from json import JSONEncoder
 from typing import AbstractSet, Any, Dict, List, Optional, Set, Type, Union
 
 import isodate
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, StrictStr
 
 from globus_action_provider_tools.utils import (
     now_isoformat,
     principal_urn_regex,
     shortish_id,
-    uuid_regex,
 )
 
 
@@ -63,6 +62,12 @@ class ActionProviderDescription(BaseModel):
     runnable_by: List[str] = Field(default_factory=lambda: ["all_authenticated_users"])
     administered_by: Optional[List[str]] = None
     event_types: Optional[List[EventType]] = None
+
+
+class RequestObject(BaseModel):
+    """Use pydantic to enforce that the request object is a Python dictionary."""
+
+    __root__: Dict[StrictStr, Any]
 
 
 class ActionRequest(BaseModel):
@@ -272,16 +277,20 @@ class ActionStatus(BaseModel):
         return self.status in (ActionStatusValue.SUCCEEDED, ActionStatusValue.FAILED)
 
 
-class ActionProviderJsonEncoder(JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, AbstractSet):
-            return list(obj)
-        elif isinstance(obj, BaseModel):
-            return obj.dict()
-        elif inspect.isclass(obj) and issubclass(obj, BaseModel):
-            return obj.schema()
-        elif isinstance(obj, datetime.datetime):
-            return obj.isoformat()
-        elif isinstance(obj, datetime.timedelta):
-            return isodate.duration_isoformat(obj)
-        return super(ActionProviderJsonEncoder, self).default(obj)
+def convert_to_json(o: Any) -> Any:
+    if isinstance(o, AbstractSet):
+        return list(o)
+    elif isinstance(o, BaseModel):
+        return o.dict()
+    elif inspect.isclass(o) and issubclass(o, BaseModel):
+        return o.schema()
+    elif isinstance(o, datetime.datetime):
+        return o.isoformat()
+    elif isinstance(o, datetime.timedelta):
+        return isodate.duration_isoformat(o)
+    return json.JSONEncoder().default(o)
+
+
+class ActionProviderJsonEncoder(json.JSONEncoder):
+    def default(self, o: Any) -> Any:
+        return convert_to_json(o)