+++ CHANGELOG.rst
@@ -10,6 +10,49 @@ Unreleased changes are documented in files in the `changelog.d`_ directory.
 
 ..  scriv-insert-here
 
+.. _changelog-0.13.0b1:
+
+0.13.0b1 — 2022-12-14
+=====================
+
+Python support
+--------------
+
+- Add support for Python 3.11.
+- Drop support for Python 3.6.
+
+Bugfixes
+--------
+
+-   Fix a crash that will occur if a non-object JSON document is submitted.
+    For example, this will happen if the incoming JSON document is ``"string"``
+    or ``["array"]``.
+
+- Fix a crash that occurs when an HTTP 400 "invalid grant" error is received
+  from Globus Auth while getting an authorizer for a given scope.
+
+  This is now caught by ``AuthState.get_authorizer_for_scope()`` and ``None`` is returned.
+
+Changes
+-------
+
+-   Remove the ``__version__`` attribute.
+
+    The ``importlib.metadata`` module in Python 3.8 and higher
+    (or the backported ``importlib_metadata`` package)
+    can be used to query the version of installed packages if needed.
+
+- ``jsonschema>=4.17,<5`` is now required by action-provider-tools.
+
+  Consumers of the library will have to update to a more recent version of ``jsonschema``
+  if they are using it explicitly.
+
+0.12.0 - 2022-03-02
+===================
+
+*No changes from 0.12.0b1.*
+
+
 0.12.0b1 - 2022-02-11
 =====================
 
@@ -62,7 +105,7 @@ Documentation
 -------------
 
 - Adds documentation around the new caching behavior:
-  https://action-provider-tools.readthedocs.io/en/latest/toolkit/caching.html 
+  https://action-provider-tools.readthedocs.io/en/latest/toolkit/caching.html
 
 
 0.11.3 - 2021-05-27
@@ -94,7 +137,7 @@ Features
 
 - Allows the detail field to be a string.
 - Improves logging output in the case where there is an Action Provider throws
-  Exceptions or an authentication issue. 
+  Exceptions or an authentication issue.
 - Allows for environment variable configuration.
 - Bundles Flask an an optional dependency. See the README.md for information on
   installing the toolkit with Flask.
@@ -112,7 +155,7 @@ Deprecations
 ------------
 
 - The Flask Callback Loader Helper is now deprecated in favor of the Flask
-  Blueprint Helper. 
+  Blueprint Helper.
 
 0.11.0 - 2021-03-29
 ===================
@@ -159,7 +202,7 @@ Features
 --------
 
 - Improves testing tools for isolating tests between different instances of
-  ActionProviderBlueprints and the Flask helpers. 
+  ActionProviderBlueprints and the Flask helpers.
 
 0.10.3 - 2020-10-01
 ===================
@@ -168,11 +211,11 @@ Features
 --------
 
 - Adds a shared patch to the testing library to mock out an
-  ActionProviderBlueprints TokenChecker 
+  ActionProviderBlueprints TokenChecker
 - Users can now specify a Globus Auth Client Name (legacy) when creating an
-  instance of the ActionProviderBlueprint 
+  instance of the ActionProviderBlueprint
 - Users can now specify multiple acceptable scopes when creating an instance of
-  the ActionProviderBlueprint 
+  the ActionProviderBlueprint
 
 Bugfixes
 --------